move all managers into own directory and mark old ones as deprecated
