+++ src/BqRestManager.php
@@ -2,78 +2,9 @@
 
 namespace Meisterwerk\BqUtils;
 
-class BqRestManager
+/**
+ * @deprecated class was moved into directory "Manager"
+ */
+class BqRestManager extends Manager\BqRestManager
 {
-    private string $apiKey;
-
-    private string $url;
-
-    public function __construct(string $apiKey, string $url)
-    {
-        $this->apiKey = $apiKey;
-        $this->url = $url;
-    }
-
-    /**
-     * @throws BqRequestException
-     */
-    public function get(string $endpoint, $jsonAssociative = false, $jsonDecode = true)
-    {
-        return BqUtil::request([
-            CURLOPT_URL => $this->url .  $endpoint,
-            CURLOPT_CUSTOMREQUEST => 'GET',
-            CURLOPT_HTTPHEADER => [
-                'Authorization: Bearer '.$this->apiKey
-            ],
-        ], $jsonAssociative, $jsonDecode);
-    }
-
-    /**
-     * @throws BqRequestException
-     */
-    public function put(string $endpoint, $postFields, $jsonAssociative = false, $jsonDecode = true)
-    {
-        return BqUtil::request([
-            CURLOPT_URL => $this->url . $endpoint,
-            CURLOPT_CUSTOMREQUEST => 'PUT',
-            CURLOPT_POSTFIELDS => json_encode($postFields),
-            CURLOPT_HTTPHEADER => [
-                'Authorization: Bearer '.$this->apiKey,
-                'Accept: application/json',
-                'Content-Type: application/json',
-            ],
-        ], $jsonAssociative, $jsonDecode);
-    }
-
-    /**
-     * @throws BqRequestException
-     */
-    public function post(string $endpoint, $postFields, $jsonAssociative = false, $jsonDecode = true)
-    {
-        return BqUtil::request([
-            CURLOPT_URL => $this->url . $endpoint,
-            CURLOPT_CUSTOMREQUEST => 'POST',
-            CURLOPT_POSTFIELDS => json_encode($postFields),
-            CURLOPT_HTTPHEADER => [
-                'Authorization: Bearer '.$this->apiKey,
-                'Accept: application/json',
-                'Content-Type: application/json',
-            ],
-        ], $jsonAssociative, $jsonDecode);
-    }
-
-    /**
-     * @throws BqRequestException
-     */
-    public function delete(string $endpoint, $jsonAssociative = false, $jsonDecode = true)
-    {
-        return BqUtil::request([
-            CURLOPT_URL => $this->url . $endpoint,
-            CURLOPT_CUSTOMREQUEST => 'DELETE',
-            CURLOPT_HTTPHEADER => [
-                'Authorization: Bearer '.$this->apiKey,
-                'Content-Type: application/json'
-            ],
-        ], $jsonAssociative, $jsonDecode);
-    }
 }
