+++ src/Manager/BqOrderCacheManager.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Meisterwerk\BqUtils\Manager;
+
+class BqOrderCacheManager
+{
+    private FileManager $fileManager;
+
+    public function __construct(string $cachePath)
+    {
+        $this->fileManager = new FileManager($cachePath);
+    }
+
+    public function getOrder($orderNumber)
+    {
+        return $this->fileManager->getFileData($orderNumber);
+    }
+
+    public function setOrder($orderData): void
+    {
+        $this->fileManager->setFileData($orderData);
+    }
+}
